Working robot with gobuilda odometry.  Lifts are working with encoders.
