+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/GoBuilda/GoBuildaDriveToPointExample.java
@@ -7,6 +7,7 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.Pose2D;
+import org.firstinspires.ftc.teamcode.Robot.TT_RobotHardware;
 
 import java.util.Locale;
 
@@ -14,14 +15,7 @@ import java.util.Locale;
 //@Disabled
 
 public class GoBuildaDriveToPointExample extends LinearOpMode {
-
-    DcMotor leftFrontDrive;
-    DcMotor rightFrontDrive;
-    DcMotor leftBackDrive;
-    DcMotor rightBackDrive;
-
-    GoBuildaPinpointDriver odo; // Declare OpMode member for the Odometry Computer
-    GoBuildaDriveToPoint nav = new GoBuildaDriveToPoint(this); //OpMode member for the point-to-point navigation class
+    TT_RobotHardware robot = new TT_RobotHardware(this);
 
     enum StateMachine {
         WAITING_FOR_START,
@@ -33,51 +27,27 @@ public class GoBuildaDriveToPointExample extends LinearOpMode {
         DRIVE_TO_TARGET_5
     }
 
-    static final Pose2D TARGET_1 = new Pose2D(DistanceUnit.MM,1000,20,AngleUnit.DEGREES,0);
-    static final Pose2D TARGET_2 = new Pose2D(DistanceUnit.MM, 1000, -20, AngleUnit.DEGREES, -90);
-    static final Pose2D TARGET_3 = new Pose2D(DistanceUnit.MM,1000,-40, AngleUnit.DEGREES,-180);
-    static final Pose2D TARGET_4 = new Pose2D(DistanceUnit.MM, 2000, -40, AngleUnit.DEGREES, -180);
-    static final Pose2D TARGET_5 = new Pose2D(DistanceUnit.MM, 2000, 0, AngleUnit.DEGREES, 0);
-
+    static final Pose2D TARGET_1 = new Pose2D(DistanceUnit.MM,500,0,AngleUnit.DEGREES,0);
+    static final Pose2D TARGET_2 = new Pose2D(DistanceUnit.MM, 0, 0, AngleUnit.DEGREES, 0);
+    static final Pose2D TARGET_3 = new Pose2D(DistanceUnit.MM,0,0, AngleUnit.DEGREES,0);
+    static final Pose2D TARGET_4 = new Pose2D(DistanceUnit.MM, 0, 0, AngleUnit.DEGREES, 0);
+    static final Pose2D TARGET_5 = new Pose2D(DistanceUnit.MM, 0, 0, AngleUnit.DEGREES, 90);
 
     @Override
     public void runOpMode() {
 
         // Initialize the hardware variables. Note that the strings used here must correspond
         // to the names assigned during the robot configuration step on the DS or RC devices.
-
-        leftFrontDrive  = hardwareMap.get(DcMotor.class, "Left Front");
-        leftBackDrive  = hardwareMap.get(DcMotor.class, "Left Back");
-        rightFrontDrive = hardwareMap.get(DcMotor.class, "Right Front");
-        rightBackDrive = hardwareMap.get(DcMotor.class, "Right Back");
-
-        leftFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        rightFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        leftBackDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        rightBackDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-
-        leftFrontDrive.setDirection(DcMotor.Direction.FORWARD);
-        leftBackDrive.setDirection(DcMotor.Direction.FORWARD);
-        rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
-        rightBackDrive.setDirection(DcMotor.Direction.REVERSE);
-
-        odo = hardwareMap.get(GoBuildaPinpointDriver.class,"odo");
-        odo.setOffsets(-142.0, 120.0); //these are tuned for 3110-0002-0001 Product Insight #1
-        odo.setEncoderResolution(GoBuildaPinpointDriver.GoBildaOdometryPods.goBILDA_4_BAR_POD);
-        odo.setEncoderDirections(GoBuildaPinpointDriver.EncoderDirection.FORWARD, GoBuildaPinpointDriver.EncoderDirection.FORWARD);
-
-        odo.resetPosAndIMU();
-
-        nav.setDriveType(GoBuildaDriveToPoint.DriveType.MECANUM);
+        robot.init();
 
         StateMachine stateMachine;
         stateMachine = StateMachine.WAITING_FOR_START;
 
         telemetry.addData("Status", "Initialized");
-        telemetry.addData("X offset", odo.getXOffset());
-        telemetry.addData("Y offset", odo.getYOffset());
-        telemetry.addData("Device Version Number:", odo.getDeviceVersion());
-        telemetry.addData("Device Scalar", odo.getYawScalar());
+        telemetry.addData("X offset", robot.odo.getXOffset());
+        telemetry.addData("Y offset", robot.odo.getYOffset());
+        telemetry.addData("Device Version Number:", robot.odo.getDeviceVersion());
+        telemetry.addData("Device Scalar", robot.odo.getYawScalar());
         telemetry.update();
 
         // Wait for the game to start (driver presses START)
@@ -85,7 +55,7 @@ public class GoBuildaDriveToPointExample extends LinearOpMode {
         resetRuntime();
 
         while (opModeIsActive()) {
-            odo.update();
+            robot.odo.update();
 
             switch (stateMachine){
                 case WAITING_FOR_START:
@@ -98,32 +68,32 @@ public class GoBuildaDriveToPointExample extends LinearOpMode {
                     the robot has reached the target, and has been there for (holdTime) seconds.
                     Once driveTo returns true, it prints a telemetry line and moves the state machine forward.
                      */
-                    if (nav.driveTo(odo.getPosition(), TARGET_1, 0.7, 0)){
+                    if (robot.nav.driveTo(robot.odo.getPosition(), TARGET_1, 0.7, 10)){
                         telemetry.addLine("at position #1!");
                         stateMachine = StateMachine.DRIVE_TO_TARGET_2;
                     }
                     break;
                 case DRIVE_TO_TARGET_2:
                     //drive to the second target
-                    if (nav.driveTo(odo.getPosition(), TARGET_2, 0.7, 1)){
+                    if (robot.nav.driveTo(robot.odo.getPosition(), TARGET_2, 0.7, 3)){
                         telemetry.addLine("at position #2!");
                         stateMachine = StateMachine.DRIVE_TO_TARGET_3;
                     }
                     break;
                 case DRIVE_TO_TARGET_3:
-                    if(nav.driveTo(odo.getPosition(), TARGET_3, 0.7, 3)){
+                    if(robot.nav.driveTo(robot.odo.getPosition(), TARGET_3, 0.7, 3)){
                         telemetry.addLine("at position #3");
                         stateMachine = StateMachine.DRIVE_TO_TARGET_4;
                     }
                     break;
                 case DRIVE_TO_TARGET_4:
-                    if(nav.driveTo(odo.getPosition(),TARGET_4,0.7,1)){
+                    if(robot.nav.driveTo(robot.odo.getPosition(),TARGET_4,0.7,3)){
                         telemetry.addLine("at position #4");
                         stateMachine = StateMachine.DRIVE_TO_TARGET_5;
                     }
                     break;
                 case DRIVE_TO_TARGET_5:
-                    if(nav.driveTo(odo.getPosition(),TARGET_5,0.7,1)){
+                    if(robot.nav.driveTo(robot.odo.getPosition(),TARGET_5,0.7,3)){
                         telemetry.addLine("There!");
                         stateMachine = StateMachine.AT_TARGET;
                     }
@@ -132,18 +102,16 @@ public class GoBuildaDriveToPointExample extends LinearOpMode {
 
 
             //nav calculates the power to set to each motor in a mecanum or tank drive. Use nav.getMotorPower to find that value.
-            leftFrontDrive.setPower(nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.LEFT_FRONT));
-            rightFrontDrive.setPower(nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.RIGHT_FRONT));
-            leftBackDrive.setPower(nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.LEFT_BACK));
-            rightBackDrive.setPower(nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.RIGHT_BACK));
+            robot.setDrivePower(robot.nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.LEFT_FRONT),robot.nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.RIGHT_FRONT),
+                    robot.nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.LEFT_BACK),robot.nav.getMotorPower(GoBuildaDriveToPoint.DriveMotor.RIGHT_BACK));
 
             telemetry.addData("current state:",stateMachine);
 
-            Pose2D pos = odo.getPosition();
+            Pose2D pos = robot.odo.getPosition();
             String data = String.format(Locale.US, "{X: %.3f, Y: %.3f, H: %.3f}", pos.getX(DistanceUnit.MM), pos.getY(DistanceUnit.MM), pos.getHeading(AngleUnit.DEGREES));
             telemetry.addData("Position", data);
-
             telemetry.update();
+
 
         }
     }}
