+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/ServoTesting.java
@@ -29,8 +29,11 @@
 
 package org.firstinspires.ftc.teamcode.Testing;
 
+import android.widget.Button;
+
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.hardware.DigitalChannel;
 import com.qualcomm.robotcore.hardware.Servo;
 
 /*
@@ -51,15 +54,17 @@ import com.qualcomm.robotcore.hardware.Servo;
 
 public class ServoTesting extends LinearOpMode {
 
-    static final double INCREMENT   = 0.01;     // amount to slew servo each CYCLE_MS cycle
-    static final int    CYCLE_MS    =   50;     // period of each cycle
+    static final double INCREMENT = 0.01;     // amount to slew servo each CYCLE_MS cycle
+    static final int CYCLE_MS = 50;     // period of each cycle
     //static final double MAX_POS     =  1.0;     // Maximum rotational position
-    static final double MAX_POS     =  0.3;
-    static final double MIN_POS     =  0.0;     // Minimum rotational position
+    static final double MAX_POS = 0.3;
+    static final double MIN_POS = 0.0;     // Minimum rotational position
 
     // Define class members
-    Servo   servo;
-    double  position = (MAX_POS - MIN_POS) / 2; // Start at halfway position
+    Servo servo;
+    DigitalChannel button;
+
+    double position = (MAX_POS - MIN_POS) / 2; // Start at halfway position
     boolean rampUp = true;
 
 
@@ -68,42 +73,46 @@ public class ServoTesting extends LinearOpMode {
 
         // Connect to servo (Assume Robot Left Hand)
         // Change the text in quotes to match any servo name on your robot.
-        servo = hardwareMap.get(Servo.class, "Lift Arm");
-        position = servo.getPosition();
+        servo = hardwareMap.get(Servo.class, "test");
+        button = hardwareMap.get(DigitalChannel.class, "button");
+        button.setMode(DigitalChannel.Mode.INPUT);
+        //if (servo.get)
+        position = 0.5; //servo.getPosition();
         // Wait for the start button
-        telemetry.addData(">", "Press Start to scan Servo." );
+        telemetry.addData(">", "Press Start to scan Servo.");
         telemetry.update();
         waitForStart();
 
 
         // Scan servo till stop pressed.
-        while(opModeIsActive()){
+        while (opModeIsActive()) {
 
-            // slew the servo, according to the rampUp (direction) variable.
-            if (gamepad1.a) {
-                // Keep stepping up until we hit the max value.
-                position += INCREMENT ;
+            if (!button.getState() || button.getState()) {
+                // slew the servo, according to the rampUp (direction) variable.
+                if (gamepad1.a) {
+                    // Keep stepping up until we hit the max value.
+                    position += INCREMENT;
+                } else if (gamepad1.b) {
+                    // Keep stepping down until we hit the min value.
+                    position -= INCREMENT;
+                }
+
+
+            } else {
+                position = 0.5;
             }
-            else  if (gamepad1.b) {
-                // Keep stepping down until we hit the min value.
-                position -= INCREMENT ;
-            }
+
+            servo.setPosition(position);
 
             // Display the current value
             telemetry.addData("Servo Position", "%5.2f", servo.getPosition());
-            telemetry.addData("Servo Target", "%5.2f", position );
-            telemetry.addData(">", "Press Stop to end test." );
+            telemetry.addData("Servo Target", "%5.2f", position);
+            telemetry.addData("Button Pressed", "%s", button.getState());
+
+            telemetry.addData(">", "Press Stop to end test.");
             telemetry.update();
-
-            // Set the servo to the new position and pause;
-            servo.setPosition(position);
-
             sleep(CYCLE_MS);
             idle();
         }
-
-        // Signal done;
-        telemetry.addData(">", "Done");
-        telemetry.update();
     }
 }
